MockNetworkManager와 Dummy Data 분리.
